Add walls
This was mostly a modification to the lighting algorithm, which used not
to take into account unwalkable but visible tiles (like walls) so rather
than fixing the issue, I removed the unwalkable but visible tiles from the
map. Anyway, turns out the fix wasn't too hard since it can make use of
existing code.